fix(electron): TabData interface moved to electron
