+++ electron/src/storage/types/local-repository.ts
@@ -1,7 +1,7 @@
 import {User} from "../../sbg-api-client/interfaces/user";
 import {ExecutorConfig} from "./executor-config";
 import {RepositoryType} from "./repository-type";
-import {TabData} from "../../../../src/app/core/workbox/tab-data.interface";
+import {TabData} from "./tab-data-interface";
 
 export interface CredentialsCache {
     id: string;
