+++ electron/src/storage/types/repository-type.ts
@@ -1,6 +1,6 @@
 import {RecentAppTab} from "./recent-app-tab";
 import {AppMeta} from "./app-meta";
-import {TabData} from "../../../../src/app/core/workbox/tab-data.interface";
+import {TabData} from "./tab-data-interface";
 
 export class RepositoryType {
     activeTab: { tabID: string; activationTime: number; };
