+++ electron/src/storage/types/tab-data-interface.ts
