+++ src/app/repository/platform-repository.service.ts
@@ -10,7 +10,7 @@ import {Project} from "../../../electron/src/sbg-api-client/interfaces/project";
 import {RawApp} from "../../../electron/src/sbg-api-client/interfaces/raw-app";
 import {AppMeta} from "../../../electron/src/storage/types/app-meta";
 import {RecentAppTab} from "../../../electron/src/storage/types/recent-app-tab";
-import {TabData} from "../core/workbox/tab-data.interface";
+import {TabData} from "../../../electron/src/storage/types/tab-data-interface";
 import {IpcService} from "../services/ipc.service";
 import {AuthService} from "../auth/auth.service";
